+++ Gamepy/.gitignore
@@ -1,3 +1,5 @@
+#randos also found
+*.idea/
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
