--- Gamepy/init
@@ -1,1 +1,0 @@
-##first file here just to make the gamepy folder to clone into.
