+++ Gamepy/run_app.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+python3 app.py ${@}
